feat: new file for shuffling superdroplets algorithm
